refactor: Add biome configuration, improve CartDrawer and Topbar components, and update D2Bot encryption methods
